requirements.txt
autogen>=0.2.0
PyMuPDF>=1.23.0
reportlab>=4.0.0
python-dotenv>=1.0.0
typer>=0.9.0
rich>=13.7.0
tqdm>=4.66.0
pytest>=7.4.0 